+++ src/OpenTelemetry/Logs/OpenTelemetryLoggerOptions.cs
@@ -17,12 +17,14 @@
 #if NET461 || NETSTANDARD2_0
 using System;
 using System.Collections.Generic;
+using OpenTelemetry.Resources;
 
 namespace OpenTelemetry.Logs
 {
     public class OpenTelemetryLoggerOptions
     {
         internal readonly List<BaseProcessor<LogRecord>> Processors = new List<BaseProcessor<LogRecord>>();
+        internal ResourceBuilder ResourceBuilder = ResourceBuilder.CreateDefault();
 
         /// <summary>
         /// Gets or sets a value indicating whether or not log scopes should be
@@ -63,6 +65,19 @@ namespace OpenTelemetry.Logs
             }
 
             this.Processors.Add(processor);
+
+            return this;
+        }
+
+        /// <summary>
+        /// Sets the <see cref="ResourceBuilder"/> from which the Resource associated with
+        /// this provider is built from. Overwrites currently set ResourceBuilder.
+        /// </summary>
+        /// <param name="resourceBuilder"><see cref="ResourceBuilder"/> from which Resource will be built.</param>
+        /// <returns>Returns <see cref="OpenTelemetryLoggerOptions"/> for chaining.</returns>
+        public OpenTelemetryLoggerOptions SetResourceBuilder(ResourceBuilder resourceBuilder)
+        {
+            this.ResourceBuilder = resourceBuilder ?? throw new ArgumentNullException(nameof(resourceBuilder));
 
             return this;
         }