+++ src/OpenTelemetry.Exporter.Console/CHANGELOG.md
@@ -11,7 +11,9 @@ please check the latest changes
 
 * Removed code that prints Baggage information
   ([#1825](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1825))
-* LogRecordExporter exports Message and StateValues from LogRecord.
+* LogRecordExporter exports Message, Scope, StateValues from LogRecord.
+  ([#1871](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1871)
+  [#1895](https://github.com/open-telemetry/opentelemetry-dotnet/pull/1895))
 
 ## 1.0.1